Bulk check for existing HACK CLUB FEE for CanonicalTransaction in events#show (#2465)
Next PR for https://github.com/hackclub/bank/issues/2446

Query/allocation count before: SQL Queries: 1223 (520 cached) | Allocations: 3611538
Query/allocation count after: SQL Queries: 1124 (520 cached) | Allocations: 3290516

From this query:
```
  Fee Exists? (0.5ms)  SELECT 1 AS one FROM "fees" INNER JOIN "canonical_event_mappings" ON "fees"."canonical_event_mapping_id" = "canonical_event_mappings"."id" WHERE "canonical_event_mappings"."canonical_transaction_id" = $1 AND "fees"."reason" = $2 LIMIT $3  [["canonical_transaction_id", redacted], ["reason", "HACK CLUB FEE"], ["LIMIT", 1]]
  ↳ app/models/canonical_transaction.rb:164:in `fee_payment?'
```

which will be an N+1 here
https://github.com/hackclub/bank/blob/7cad289fe458b6fd63512c1ff1d7db8f70409264/app/models/canonical_transaction.rb#L164

Also fixed a subtle problem with memoizing `@fee_payment`. Since it stores
a boolean, if the value is `false`, it doesn't get cached with `||=` and
instead runs the `exists?` query again. So use a guard return with
`defined?` instead.

It might be worth fixing this in other places with memoization as a defensive
measure (`||=` will also not cache `nil`s properly), but that should be a separate task.