Allow Users to choose Session length (#2504)
- Add `session_duration_seconds` column to `users` with a default of 30 days (to replace the hardcoded constant)
- Add `expiration_at` to `user_sessions` to track when a session should be expired
- Set `expiration_at` for existing (both deleted and non deleted) sessions to +30 days from `created_at`
- Use a user's `session_duration_seconds` to compute `expiration_at` when creating a new session
- Use `expiration_at` when checking for session expiration instead of hardcoded constant

`session_duration_seconds` might belong on a separate table from `users` but it was the easiest place to add it for now. This also means the easiest place to put the frontend for setting was with the rest of the user edit form, but it's possible it might make sense lower on the page under Login Details (maybe in the Sessions section?). One drawback of moving it there would be that it could complicate the UX of the form (2 separate form sections on the page with separate submits, or a modal like the security key). Deferring this decision for later