Create user on invite if they don't exist (#2650)
* Create user on invite if they don't exist

https://github.com/hackclub/bank/issues/2415

Previously, if a non existing user was invited, we would not associate
user to OrganizerPositionInvite until the user logged in. This led to
some complex logic trying to dynamically associate the user later, see

- https://github.com/hackclub/bank/pull/2327
- https://github.com/hackclub/bank/pull/2473

While it's not entirely clear why this approach didn't work, we can just
circumvent it by creating the user when creating the invite. This
doesn't break any existing assumptions about users (i.e. the presence of
a user record does NOT imply that the user has signed in).

For implementing this change, I could have tried using callbacks on the
`OrganizerPositionInvite` model to create the user, but it is generally a
bad idea to have callbacks touch another model as it tends to get
messy quickly (see https://canolcer.com/post/active-record-callbacks/)

So I went with a service object to follow the convention of the rest of
the codebase. This had a few consequences:

- the service still has to expose the model for use with form_with and
  some error rendering. I will explore whether we can avoid passing the
  model to `form_with` and if we can render errors directly from the service
  in a follow up.
- we were calling the ! version of persisting (create/save) in some
places and not in others, so I added `run` vs `run!`
- I moved the `normalize_email` method from the model to the service
  since we need it to happen before creating the user, using safe operator for when email is empty on the new action

* Add dependent destroy otherwise there are FK errors when deleting a user with associated organizer_position_invite

* Add test for OrganizerPositionInviteService::Create

I didn't use factories yet for this test because some of the setup for
that is in another pending PR, so I'd rather wait for that to land and
then refactor this one.