Unskip g_suite_service/mark_verified_spec.rb and replace fixtures with factories (#3045)
https://github.com/hackclub/bank/issues/3008

This set of tests actually passed without doing anything, so not sure why it was skipped. I replaced fixtures with factories and improved the context organization. Frequently tests end up being in a 2D grid, such as

|| verified_on_google == true      | verified_on_google == false |
|-| ----------- | ----------- |
| no created_by|changes state, no email      | no change state, no email       |
| has created_by|sends email   | no change state, no email        |

but code ends up being more "1D", so I find it helps to pick one axis as the top level one and arrange all tests that way. In this PR I chose the `verified_on_google` value

* Unskip GSuiteService::MarkVerified spec and everything passes

* Replace fixtures with factories

* Rearrange tests into better contexts.

The top level thing we care about in this service is whether
verified_on_google is true or false, so that should be the top level
context. Then we can have a nested context for each case that checks
behavior when created_by exists