Replace StripeCardholders table in card_overview with a button to view cardholder in stripe (#3042)
Closes #2554

The card_overview page (e.g. https://bank.hackclub.com/hq/cards) was very slow for admin users because there was a cardholders table on the bottom that would require N calls to stripe to show whether the cardholder was active. Per the conversation on #2554 (comment), this table wasn't very useful for admin users, so instead remove that table and add a button per card to view cardholders directly on stripe.