Unskip g_suite_mailer_spec and convert fixtures to factories (#3067)
https://github.com/hackclub/bank/issues/3008

The main issue here is that the URL being checked didn't conform to the actual structure for the event's gsuite overview. With factories we have to make it more dynamic, but because the implementation is using the event_g_suite_overview_url helper, it'd be kind of a tautological test, so I checked for an expected URL built with File.join instead.


* Unskip g_suite_mailer_spec and fix

It looks like we are looking for the gsuite event URL, which is not the
URL that was in the test

* Convert fixtures to factories

This required updating the URL to be more dynamic. We could have used
the event_g_suite_overview_url helper but that's the same method the
implementation uses so it'd be kind of a tautological test, so I checked
for a URL built with File.join instead.