Associate HashedTransaction with CanonicalTransaction during create, not afterwards (part 2) (#3104)
https://github.com/hackclub/bank/issues/3103

Turns out there are 3 other import services that need similar fixes to https://github.com/hackclub/bank/pull/3076

The first 2 were straightforward (`EmburseThatLookLikeDuplicates` and `StripeThatLookLikeDuplicates`). They seem to ignore the "potentially duplicate" issue and just create `CanonicalTransaction`s the same as the `Simple` service does.

The third one `PlaidThatLookLikeDuplicates` looks like it was intending to skip certain duplicate transactions but I believe it was not functioning as intended and was skipping processing both hashed transactions if they were considered duplicates, instead of processing one and skipping the other. This has now been fixed.

* Fix logic for checking duplicate plaid transactions

We have a duplicate if there's another HashedTransaction with the same
primary_hash, but a different raw_plaid_transaction_id. We should import one of them but skip the other.