Unskip auth_service/token_spec and convert fixtures to factories (#3068)
* Unskip and fix auth_service/token_spec

Just unskipping this test by itself would fail the 2 checks for
UnauthenticatedError when login_token didn't exist, which is exactly
what this begin-rescue-ensure block seems to be trying to implement, so
I was confused for a while.

It turns out if login_token didn't exist, then the rescue would run, but
then before the `raise UnauthenticatedError`, the ensure block would try
to run and throw `ActiveRecord::RecordNotFound` yet again when doing the
`used?` check. It seems like there's no point in running the code in the
ensure block if the login_token doesn't already exist, so I moved it
into the normal method body before returning user.

I also wanted to check that user (which runs login_token.user) wouldn't
also possibly be a culprit for `ActiveRecord::RecordNotFound`, but
because `login_tokens` has a FK constraint on the db with `users`, this
should never happen.

I ended up removing the inline rescue because it only seems to be
catching an invalid AASM transition error which I wasn't sure if it made
sense to do. The scary thing is that inline rescue can apparently catch
a lot of errors that we usually didn't intendo to. Reference:
https://thoughtbot.com/blog/don-t-inline-rescue-in-ruby

* Create login_token_factory and use for login_token_spec

* Replace fixtures with factories in auth_service/token_spec

* Skip `mark_used` transition if already_used

Co-authored-by: Gary Tou <gary@garytou.com>