Show running balance when ?show_running_balance=true (#3757)
- Restructure call to Kaminari.paginate_array so we can reuse some of the variables

It turns out the way we are paginating transactions here does not
actually decrease the number of `CanonicalTransactionGrouped` instances we query for.
While this is a performance improvement we should make, it turned out to
be convenient for reusing here to compute the running balance

Since we already query for all the transactions when showing an
organization (even though the UI is paginated), the general approach is
sum up the amounts of every transaction that comes on the next page up
to the last (i.e. every transaction before the oldest one on the current
page). Then we add each transaction's amount per row cumulatively.

Currently we only show the running balance if we pass the query param
`show_running_balance=true` and the currently logged in user is an
admin. This triggers a huge N+1 because the `amount` method looks up
`canonical_transactions` and `hcb_code` per
`CanonicalTransactionGrouped`, but as the issue stated, I didn't do any
performance improvements for this first pass. Because we are still using
`CanonicalTransactionGrouped`, we will probably have to use a similar
strategy to the preloader classes.