SMS Users When IRL Transactions Are Declined (#4158)
This PR also clarifies that `send_twilio_message` was for receipts by
renaming it to `send_twilio_receipt_message`. I'm not sure if it's best
to use the Flipper flag for receipts in this case, however, I left it
for now. Will wait to hear what others think. I think this fixes #3560.

---------

Co-authored-by: yodalightsabr <yoda@yodacode.xyz>