adds grid view to mock cards (#4636)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
onboarding would like grid view to be default when demoing the platform.
grid view is currently broken.


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
this fixes it by adding keys to fix render and adding a to_partial_path
key to help rails find the partial due to the openstruct not having the
model struct. also fixes card muted status and href on cards.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->