Generate `BankFee` memo from HCB code (#4947)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->

Currently, the "`FISCAL SPONSORSHIP`" memo for fees is being generated
at the CT level instead of the HCB code level. This is inconsistent with
how other default memos are generated.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->

now it's generated by the HCB code B)

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->