Fix Double Click To Rename on `/my/inbox` (#5033)
## Summary of the problem

If there was a suggested pairing for a transaction on `/my/inbox`, we'd
render the transaction twice (once in the pairing carousel and once in
the list). This meant that we'd also render two Turbo frames with the
same ID. That meant that when Turbo went to replace the frame with the
rename form it would replace the wrong one.

## Describe your changes

I've disabled rename on transaction when they're displayed in the
pairing context. Now we only render one Turbo Frame!