Add anonymous donations (#5141)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
Some donors would like to remain anonymous when donating to a
transparent organization.
Closes #5112 


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
This PR adds a checkbox on the start donation page of an event that
allows donors to choose to be anonymous. A tooltip explains the feature.
If checked, the donation will have the new `anonymous` field on the
donation schema set to true. Anonymity is reflected across the ledger
(with a custom memo), the donations page, and the transparency API. This
checkbox is also behind a feature preview for each event.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

![image](https://github.com/hackclub/hcb/assets/79737178/5ac9682a-b025-44e6-990a-62aec72ef531)

---------

Co-authored-by: yodalightsabr <yoda@yodacode.xyz>
Co-authored-by: Sam Poder <sam.r.poder@gmail.com>
Co-authored-by: Gary Tou <gary@garytou.com>