Standardize card colors and add new US-only limitation copy (#5187)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
Colors for the "important info" cards were mismatched across multiple
pages, and copy about limitations for US-only features was missing from
them.
Closes #5153
Closes #5157 

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Adds the copy in #5153 and standardizes the colors to use orange for the
header text and gray for the help message text.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->
The new color standard (as well as additional copy for ACHs):

![image](https://github.com/hackclub/hcb/assets/79737178/414c44a3-1ce2-4d37-b670-d779e4d4a6e0)

---------

Co-authored-by: Gary Tou <gary@garytou.com>
Co-authored-by: yodalightsabr <yoda@yodacode.xyz>