Reply to Comments via Email (#5382)
Life is hard... nowadays there are so many social media apps... Slack,
iMessage, WhatsApp, Instagram, BONK, (New) Expensify, and (of course)
HCB... responding to messages on all these different platforms is tough.
We like our users, so we won't put them through the hassle of logging in
if they want to quickly respond to a comment on their phone. Instead,
we'll just let them email us their reply. And we'll handle the rest!

I based this off of GitHub's implementation, a couple of notes:

- There are no email responses to indicate success etc. GitHub does this
and I feel like it keeps the thread clean / natural.
- Currently we rely on comment IDs to identify commentable. If a comment
is deleted, your response won't be matched to a commentable. I'm alright
with this don't currently support deleting comments in the UI. If anyone
has any alternative approaches, let me know. I wanted something that was
flexible for every commentable.

---------

Co-authored-by: Gary Tou <gary@garytou.com>