Use `@inbound_mail&.mail&.subject` and then have a fallback (#5292)
This makes progress towards fixing
https://github.com/hackclub/hcb/issues/5286. I think that one issue
might be us adding a new subject (this shouldn't break it... the
`In-Reply-To` header should just work! but no Gmail!) so I've made it so
that we only use the custom subjects as fallbacks.

@cjdenio thoughts? Was there a reason you avoided doing this in
https://github.com/hackclub/hcb/pull/4786.

Also, does anyone know, is there a cleaner way to do this?