[UI] Receipts: Centralize import methods (#7662)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->

Before, the Receipt Center page had a bunch of grey text with
instructions on how to import receipts. Copying the links required using
context menus.
![CleanShot 2024-08-22 at 18 27
23@2x](https://github.com/user-attachments/assets/79cb4dae-c08c-4e05-abd8-0a8759620f8f)

For email, it sent you to your account settings to find or change your
forwarding address, which then linked back to the Receipt Bin page (we
refer to this page as Receipts, Receipt Center, Receipt Bin, and the URL
is My Inbox).
![CleanShot 2024-08-22 at 18 27
35@2x](https://github.com/user-attachments/assets/44c366b3-0b72-4cce-9928-a3eb5844812e)

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->

I've centralized these instructions and configuration as parallel
options at the top of the Receipts page:

![CleanShot 2024-08-22 at 18 28
57@2x](https://github.com/user-attachments/assets/24843585-d4fe-4d5b-848b-e9d8b62f95e6)



<!-- If there are any visual changes, please attach images, videos, or
gifs. -->