[G Suite] Fix verify button for normal users (#8053)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->

The verify button's URL used the event's ID in the slug. For security
reasons, normal users can not access events or event pages using an ID.
Instead, they must always use the slug.

We had preserved the ability for admins to access events by id. So, this
button worked for admins, but was broken for normal users. Normal users
would get a flash message "We couldn't find that organization!"

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->

I switched it to use the event's slug

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->