HQ weekly direct report transaction summary (#8647)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->

HQ needs more observability around spending. @malted's "tell zach"
mailer gets us 80% of the way there.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->

I revamped the old mailer to include missing receipts and have it sent
it to every "layer" of the organization structure. In other words, every
manager will receive one of these reports.

At the moment, I only have this set up for HCB (and graham) so we can
see if it works well. Graham is included because he requested it. We can
implement this for the entirety of Hack Club later — at that point, we
should probably fully build this into HCB.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->