Improve codespaces instructions (#8676)
## Summary of the problem
The previous instructions informed users to use docker on codespaces.
However, docker is not preinstalled and directly running `./bin/dev` is
much easier.



## Describe your changes
I changed the instructions to inform users to run `./bin/dev` instead of
continuing on to follow the docker instructions. I also included tips on
how to fix the port forwarding if it breaks.

---------

Co-authored-by: Sam Poder <sam.r.poder@gmail.com>