Use Stripe topups for fee reimbursements (#8871)
Closes https://github.com/hackclub/hcb/issues/8475

Prior to Monday 9th December. We processed fee reimbursements by making
a book transfer to cover the difference between the payout from Stripe
and the amount paid.

Now, we payout the full amount from Stripe (incl. the fee) and then top
up our Stripe balance to cover that fee.

The `FeeReimbursementService` performs that top-up.

We made this change to handle $1 payouts.