[Sidekiq] Don't explicitly set concurrency (#8878)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->

![image](https://github.com/user-attachments/assets/6f4772a6-20b9-4e94-8005-11a0664ef06a)

I belive this is because our concurrency for Sidekiq was set to 10, but
`RAILS_MAX_THREADS` was 6. this meant there were not enough database
connections to go around to the threads when there are lots of
long-running jobs queued.


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->

This allows it fallback to `RAILS_MAX_THREADS` which is better since
`RAILS_MAX_THREADS` also controls database connection pool size.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->