[Organizer Position] Drop Faker generated emails from cards (#8921)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->

People mistake it for a real email address and raise it as a security
report. We've had it happen multiple times now. Also, I've been wanting
to get rid of it for a while.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->

I'm just using `"user@example.com"` — ideally, we redesign the card so
that we don't need to show a fake email.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->