Try installing Poppler for Rubocop (#9533)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->

We've been finding that Poppler gem installs are failing on CI because
the cario native gem extension fails to build.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->

We found that Rubocop is the only job on CI that doesn't install the
poppler dependency. Maybe that's why??? Not sure

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->