[`production.Docker`] Add `.build-timestamp` to Docker image (#9559)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->

We lost the "Build x from y minutes ago" when switching to Docker. This
brings it back

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->

Write a timestamp to the Docker image at the end of the build process.
Rails then loads the timestamp once (caches via a global variable).

Global variables are generally discouraged, but I think it's a good use
case here.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->