Home page enhancements (#9698)
This pull request includes several changes to the `events_controller`
and `breakdown_engine` services to improve the handling of tags and
users, as well as updates to the corresponding view.

# TL;DR
* Merchant and user charts are now limited to a certain number of bars.
"Other" category is also removed
* Fixed small visual bugs (margin/padding/width)
* Better formatting for readability

### Improvements to tag and user handling:

*
[`app/controllers/events_controller.rb`](diffhunk://#diff-e5ba780d6ceaadee8d430a3b77cd3d401a7f0aeacec1ee78d7aa073f027af8c8L135-R136):
Corrected the condition for setting `@empty_tags` to check if `@tags` is
empty instead of `@users`.

*
[`app/services/breakdown_engine/merchants.rb`](diffhunk://#diff-2bc49cf99857fd6c9c4ccad43cec8195dec19bb3c7e8b9204405808e0c42c1dcL33-L60):
Simplified the merchant processing by sorting merchants by value in
descending order and limiting the list to the top 7 merchants.

* `app/services/breakdown_engine/users.rb`: 
- Changed the initialization of users to store them in a variable for
further processing.
- Simplified the user processing by sorting users by amount in
descending order and limiting the list to the top 11 users.

### View updates:

* `app/views/events/home/_tags_users.html.erb`: 
  - Added a `max-w-sm` class to the paragraph for better text wrapping.
- Added a bottom margin to the container of top users to improve
spacing.

Co-authored-by: Gary Tou <gary@garytou.com>