[v4 API][Card Grant] Show allowed merchants/categories (#9719)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->

allowed merchants considers the org's card grant settings. it's be nice
to show them.

merchant lock and category lock is specific to the card.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->

show allowed merchants and allowed categories. refactor the jbuilder
partial to DRY it up.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->