[Error Reporting] Fix undefined variable bug (#9818)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->

`error` is undefined. it's actually `e`.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->

I was gonna change `error` to `e`, but then I realized that we're just
printing the error message. that's not necessary since we can get the
error message from just the error object in AppSignal.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->