[Transfer wizard] Add learn more links & enhance UI (#9757)
## Summary of the problem

![image](https://github.com/user-attachments/assets/9ed01f8f-5099-466a-8bf2-fada8860a5cc)



## Describe your changes
This pull request includes changes to the `transfer_form_controller.js`
and the corresponding view template to enhance the user interface and
provide more detailed information for users. The most important changes
include updating the answer targets, modifying the question-answer logic
to include help links, and improving the visual presentation of the
form.

Enhancements to the user interface and information:

*
[`app/javascript/controllers/transfer_form_controller.js`](diffhunk://#diff-6a1a9639e264bc27b024379f02b4797eb957b7ab0ce22c705c47bee17a9511c7L16-R16):
Replaced the `helperText` target with `learnMore` to provide a link for
additional information.
*
[`app/javascript/controllers/transfer_form_controller.js`](diffhunk://#diff-6a1a9639e264bc27b024379f02b4797eb957b7ab0ce22c705c47bee17a9511c7L37-R44):
Updated the question-answer logic to include URLs for additional help
information for each answer option.
*
[`app/javascript/controllers/transfer_form_controller.js`](diffhunk://#diff-6a1a9639e264bc27b024379f02b4797eb957b7ab0ce22c705c47bee17a9511c7L74-R76):
Modified the payload handling to set the `learnMore` link dynamically
based on the selected answer.

Visual presentation improvements:

*
[`app/views/events/_transfer_form.html.erb`](diffhunk://#diff-b4c381b87f4d3901c4fd6df839be6718378f72c2016e62d610d566f243d1026fL45-R70):
Enhanced the "Help me decide" link text to be more visually appealing
and added an arrow symbol.
*
[`app/views/events/_transfer_form.html.erb`](diffhunk://#diff-b4c381b87f4d3901c4fd6df839be6718378f72c2016e62d610d566f243d1026fL45-R70):
Updated the button styles and text to improve clarity and user
experience, including changing "Exit" to "Our recommendation" and adding
a "Learn more" link.



## Screenshots


![image](https://github.com/user-attachments/assets/1f44beaf-1e5c-481f-aef8-d6a1e05f7180)

![image](https://github.com/user-attachments/assets/2066b117-04f3-43f6-82e8-7144e8c42735)


cc @sampoder

---------

Co-authored-by: Gary Tou <gary@garytou.com>