Upgrade codespaces Ruby version (#9979)
This pull request includes updates to the Ruby version used in the
development container configuration. The changes involve updating the
base image and build arguments to use a newer version of Ruby.

Updates to Ruby version:

*
[`.devcontainer/Dockerfile`](diffhunk://#diff-13bd9d7a30bf46656bc81f1ad5b908a627f9247be3f7d76df862b0578b534fc6L1-R1):
Updated the `VARIANT` argument from "3.3-bookworm" to "3.4-bookworm" to
use a newer Ruby version.
*
[`.devcontainer/devcontainer.json`](diffhunk://#diff-24ad71c8613ddcf6fd23818cb3bb477a1fb6d83af4550b0bad43099813088686L22-R22):
Updated the `VARIANT` argument from "3.3-bookworm" to "3.4.2-bookworm"
to specify the exact newer Ruby version.