[v4 API][Organization] Add Total Spent & Total Raised (#10131)
## Summary of the problem

@zachlatta would like a programmatic way to determine the total amount
spent for an organization.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->

Adding the following keys to the Organization object in the v4 API:
- `total_spent_cents`
- `total_raised_cents`

In order for these keys to returned, you must expand `reporting`.
Example:
```
http://localhost:3000/api/v4/organizations/hq?expand=reporting
```

It's important to note that both `total_spent_cents` and
`total_raised_cents` are crude calculations. It's just summing
transactions based on their amount cents. This means that a card charge
refund would be included in `total_raised_cents` and a donation refund
would be included in `total_spent_cents`. In an ideal world, we'd handle
those edge cases. However, we don't have code that does that at the
moment.