Don't include OS version in fingerprint if it doesn't exist (#10159)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
On Linux, there isn't an OS version included in the user agent string.
This caused logins from Linux machines to show as coming from "Linux
undefined".


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
`fingerprint.js` now only includes the OS version in the fingerprint if
it is defined.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->