Properly namespace jobs with aliases (#10177)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
Jobs are currently incorrectly namespaced!

Closes #8317


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Went through the incorrect jobs and renamed them to fit the rules. Also
added aliases to the old names in case they are currently in Redis -
these should be able to be removed in a month when the old names are
gone, as per the [Sidekiq
FAQ](https://github.com/sidekiq/sidekiq/wiki/FAQ#how-do-i-safely-rename-a-worker)


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->