Reduce negative margin on home page merchants graph (#10257)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
Negative margin was added to the home page merchants graph to avoid a
lot of white space on the left, added by the graph library - but this
cut off some of the labels when dealing with larger amounts:

![image](https://github.com/user-attachments/assets/a1a41946-572c-4d7e-8d0c-e426a1ce2fd1)



## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->

Reduces the negative margin to avoid cutting off larger values on the
merchants graph