Fix remaining memo ellipsis bugs (#10385)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
@Luke-Oldenburg found there was still an issue with ellipsis appearing
in memos when they shouldn't:

![image](https://github.com/user-attachments/assets/2f7b93e2-6c06-4075-a803-e2c5073ce612)



## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
The memo text is now an inline block so that the width can be set
correctly. This also meant the underline for links has to be set
manually on the span containing the text.


![image](https://github.com/user-attachments/assets/fa2d1cdb-218e-4e6f-9d03-0b307c2aecc7)




<!-- If there are any visual changes, please attach images, videos, or
gifs. -->